% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{definitions}
\alias{definitions}
\title{Relevant definitions and forest facts}
\usage{
definitions(term = "auto")
}
\arguments{
\item{term}{character, the term to be explained. It must be one of the words from this list: forest, deforestation, degradation, afforestation, reforestation, bamboo, redwoods, mangroves, importance, floods, droughts, deforestation rate, temperature.}
}
\value{
A forest fact or definition of forest concepts
}
\description{
This function provides some relevant definitions in REDD context and forest facts collected from un-redd.org. These are basic concepts that may change between the countries or projects.
}
