% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REDDworld.R
\name{REDDworld}
\alias{REDDworld}
\title{REDD worldwide 2015}
\usage{
REDDworld(variable = "AUTO")
}
\arguments{
\item{variable}{character, the information you would like to see: "forestcover" or "emissions", or leave empty to see the countries with REDD+ projects}
}
\value{
A map of countries of REDD+ projects
}
\description{
Shows a map of the situation of forest cover or emissions in the countries of REDD+ projects with data from  CIFOR– CEC – CIRAD – IFRI, 2015 (http://www.reddprojectsdatabase.org/view/countries.php)
}
